added whitelist for wallet and router addresses from tax.
This allows both DEX and CEX exchange routers to have a whitelisted wallet address that is not taxed. This is necessary as well in the event that the contract/token is bridge between different blockchains. There might be wallet and router addresses that need to be added to the whitelist.